+++ BayarBerapa/RekamTransaksiController.swift
@@ -36,27 +36,6 @@ class RekamTransaksiController: UIViewController, UIPickerViewDataSource, UIPick
         servicePicker.delegate = self
         servicePicker.dataSource = self
         
-        let headers = [
-            "Ocp-Apim-Subscription-Key": "54ff93009ab8459f88379c0203f1fccd"
-        ]
-        Alamofire.request(
-            .GET,
-            "https://program06.azure-api.net/BayarBerapaAPI/Provinsi",
-            headers: headers,
-            encoding: .JSON
-            ).responseJSON { response in
-                debugPrint(response)
-                switch response.result {
-                case .Success:
-                    let json = JSON(response.result.value!)
-                    if json != nil {
-                        self.provinceList = json.array!
-                    }
-                    break
-                case .Failure:
-                    break
-                }
-        }
     }
     
     override func didReceiveMemoryWarning() {
@@ -71,7 +50,8 @@ class RekamTransaksiController: UIViewController, UIPickerViewDataSource, UIPick
     // The number of rows of data
     func pickerView(pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
         if pickerView == provincePicker {
-            return provinceList.count
+
+            return AppManager.sharedInstance.provinceList.count
         }
         else if pickerView == pemdaPicker {
             return pemdaList.count
@@ -82,7 +62,8 @@ class RekamTransaksiController: UIViewController, UIPickerViewDataSource, UIPick
     // The data to return for the row and component (column) that's being passed in
     func pickerView(pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
         if pickerView == provincePicker {
-            
+            let json = AppManager.sharedInstance.provinceList[row]
+            return  json["NamaProvinsi"].stringValue
         }
         else if pickerView == pemdaPicker {
